refactor(chat): extract speech-to-text handling from ChatScreen
Move the device/server STT launchers, permission flow, and the engine/
language mapping helpers into rememberChatStartRecording in a new
ChatSpeechToText.kt. ChatScreen now calls the remember-helper for the
mic action instead of holding the Android intent plumbing inline.
